# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "经济法知识图谱与上市公司治理合规推理系统 API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/laws": {
      "get": {
        "tags": [
          "laws"
        ],
        "summary": "List Laws",
        "operationId": "list_laws_api_v1_laws_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/LawOut"
                  },
                  "type": "array",
                  "title": "Response List Laws Api V1 Laws Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/laws/{law_code}/versions": {
      "get": {
        "tags": [
          "laws"
        ],
        "summary": "List Law Versions",
        "operationId": "list_law_versions_api_v1_laws__law_code__versions_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "law_code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Law Code"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LegalVersionOut"
                  },
                  "title": "Response List Law Versions Api V1 Laws  Law Code  Versions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/laws/{law_code}/versions/{version_name}/articles": {
      "get": {
        "tags": [
          "laws"
        ],
        "summary": "List Law Version Articles",
        "operationId": "list_law_version_articles_api_v1_laws__law_code__versions__version_name__articles_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "law_code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Law Code"
            }
          },
          {
            "name": "version_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Version Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ArticleSummaryOut"
                  },
                  "title": "Response List Law Version Articles Api V1 Laws  Law Code  Versions  Version Name  Articles Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/laws/{law_code}/versions/{version_name}/articles/{article_no}": {
      "get": {
        "tags": [
          "laws"
        ],
        "summary": "Get Article Version",
        "operationId": "get_article_version_api_v1_laws__law_code__versions__version_name__articles__article_no__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "law_code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Law Code"
            }
          },
          {
            "name": "version_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Version Name"
            }
          },
          {
            "name": "article_no",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Article No"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArticleVersionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/laws/{law_code}/versions/{version_name}/articles/{article_no}/navigation": {
      "get": {
        "tags": [
          "laws"
        ],
        "summary": "Get Article Navigation",
        "operationId": "get_article_navigation_api_v1_laws__law_code__versions__version_name__articles__article_no__navigation_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "law_code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Law Code"
            }
          },
          {
            "name": "version_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Version Name"
            }
          },
          {
            "name": "article_no",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Article No"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArticleNavigationOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/laws/{law_code}/articles/{article_no}/effective": {
      "get": {
        "tags": [
          "laws"
        ],
        "summary": "Get Effective Article Version",
        "description": "Convenience 'as-of-date' lookup — defaults to today, matching the\nreading loop's default of showing the currently-effective text.",
        "operationId": "get_effective_article_version_api_v1_laws__law_code__articles__article_no__effective_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "law_code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Law Code"
            }
          },
          {
            "name": "article_no",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Article No"
            }
          },
          {
            "name": "at",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "At"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArticleVersionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ArticleNavigationOut": {
        "properties": {
          "current": {
            "$ref": "#/components/schemas/ArticleVersionOut"
          },
          "previous_article_no": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Previous Article No"
          },
          "next_article_no": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Article No"
          }
        },
        "type": "object",
        "required": [
          "current",
          "previous_article_no",
          "next_article_no"
        ],
        "title": "ArticleNavigationOut"
      },
      "ArticleSummaryOut": {
        "properties": {
          "article_no": {
            "type": "string",
            "title": "Article No"
          },
          "chapter_no": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chapter No"
          },
          "section_no": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Section No"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          }
        },
        "type": "object",
        "required": [
          "article_no",
          "chapter_no",
          "section_no",
          "summary"
        ],
        "title": "ArticleSummaryOut"
      },
      "ArticleVersionOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "article_id": {
            "type": "string",
            "format": "uuid",
            "title": "Article Id"
          },
          "article_no": {
            "type": "string",
            "title": "Article No"
          },
          "legal_version_id": {
            "type": "string",
            "format": "uuid",
            "title": "Legal Version Id"
          },
          "chapter_no": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chapter No"
          },
          "section_no": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Section No"
          },
          "article_text": {
            "type": "string",
            "title": "Article Text"
          },
          "valid_from": {
            "type": "string",
            "format": "date",
            "title": "Valid From"
          },
          "valid_to": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valid To"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "article_id",
          "article_no",
          "legal_version_id",
          "chapter_no",
          "section_no",
          "article_text",
          "valid_from",
          "valid_to",
          "created_at"
        ],
        "title": "ArticleVersionOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LawOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "document_type": {
            "type": "string",
            "title": "Document Type"
          },
          "issuer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Issuer"
          },
          "jurisdiction": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Jurisdiction"
          },
          "versions": {
            "items": {
              "$ref": "#/components/schemas/LegalVersionOut"
            },
            "type": "array",
            "title": "Versions"
          }
        },
        "type": "object",
        "required": [
          "id",
          "code",
          "name",
          "document_type",
          "issuer",
          "jurisdiction",
          "versions"
        ],
        "title": "LawOut"
      },
      "LegalVersionOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "version_name": {
            "type": "string",
            "title": "Version Name"
          },
          "promulgated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Promulgated At"
          },
          "effective_from": {
            "type": "string",
            "format": "date",
            "title": "Effective From"
          },
          "effective_to": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effective To"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "id",
          "version_name",
          "promulgated_at",
          "effective_from",
          "effective_to",
          "status"
        ],
        "title": "LegalVersionOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}